requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.0